requirements.txt
numpy>=1.15
Pillow>=8.0
scipy>=1.4
ruamel.yaml>=0.15
matplotlib>=3.3.0